requirements.txt
asgiref==3.2.3
attrs==19.3.0
Babel==2.9.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
Django==3.0.3
django-cors-middleware==1.5.0
django-extra-views==0.13.0
django-haystack==3.0
django-oauth-toolkit==1.2.0
django-oscar==3.0.2
# django-oscar-api==2.1.1
django-phonenumber-field==3.0.1
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==0.9.1
idna==2.8
importlib-metadata==1.5.0
mixer==6.1.3
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.1
phonenumbers==8.12.19
Pillow==7.0.0
pluggy==0.13.1
psycopg2==2.8.4
purl==1.5
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.8
wcwidth==0.1.8
zipp==3.0.0
